+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+* feat: Show `name` in `canister status` command
+* feat: `icp canister metadata <canister> <metadata section>` now fetches metadata sections from specified canisters
+* fix: Validate explicit canister paths and throw an error if `canister.yaml` is not found
+
 # v0.1.0-beta.3
 
 * feat: Remove requirement that the user install `icp-cli-network-launcher`, auto-install it on first use